DiscountedLineItemPrice as interface #1053
