+++ commercetools-models/src/main/java/io/sphere/sdk/cartdiscounts/DiscountedLineItemPrice.java
@@ -1,30 +1,17 @@
 package io.sphere.sdk.cartdiscounts;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.sphere.sdk.models.Base;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 import javax.money.MonetaryAmount;
 import java.util.List;
 
-public final class DiscountedLineItemPrice extends Base {
-    private final MonetaryAmount value;
-    private final List<DiscountedLineItemPortion> includedDiscounts;
-
-    @JsonCreator
-    private DiscountedLineItemPrice(final MonetaryAmount value, final List<DiscountedLineItemPortion> includedDiscounts) {
-        this.includedDiscounts = includedDiscounts;
-        this.value = value;
+@JsonDeserialize(as = DiscountedLineItemPriceImpl.class)
+public interface DiscountedLineItemPrice {
+    static DiscountedLineItemPrice of(final MonetaryAmount money, final List<DiscountedLineItemPortion> includedDiscounts) {
+        return new DiscountedLineItemPriceImpl(money, includedDiscounts);
     }
 
-    public static DiscountedLineItemPrice of(final MonetaryAmount money, final List<DiscountedLineItemPortion> includedDiscounts) {
-        return new DiscountedLineItemPrice(money, includedDiscounts);
-    }
+    List<DiscountedLineItemPortion> getIncludedDiscounts();
 
-    public List<DiscountedLineItemPortion> getIncludedDiscounts() {
-        return includedDiscounts;
-    }
-
-    public MonetaryAmount getValue() {
-        return value;
-    }
+    MonetaryAmount getValue();
 }
